консольная программа, создание билета/requirements.txt
Pillow==7.0.0
pip==20.0.2
pygame==1.9.6
